Added protection, this should fix bug#15182.
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@76397 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/Eiffel/interface/new_graphical/shared/editor_token_utility/eb_grid_editor_item.e
@@ -34,14 +34,16 @@ feature -- Action
 	deactivate is
 			-- Cleanup from previous call to activate.
 		do
-			Precursor {EB_GRID_EDITOR_TOKEN_ITEM}
-			if editor /= Void then
-				editor.recycle
-				editor := Void
-			end
-			if viewport /= Void then
-				viewport.destroy
-				viewport := Void
+			if not is_destroyed and then is_parented then
+				Precursor {EB_GRID_EDITOR_TOKEN_ITEM}
+				if editor /= Void then
+					editor.recycle
+					editor := Void
+				end
+				if viewport /= Void then
+					viewport.destroy
+					viewport := Void
+				end
 			end
 		end
